\name{example}
\alias{example}
\title{Example dataset}
\description{
Example dataset for GMMAT.
}
\format{
  Contains the following objects:
  \describe{
    \item{pheno}{a data frame of 400 observations on 4 variables: disease, trait, age and sex.}
    \item{GRM}{a genetic relationship matrix for 400 observations.}
  }
}


