% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_tte_data.R
\name{simulate_tte_data}
\alias{simulate_tte_data}
\title{Simulate time to event data}
\usage{
simulate_tte_data(
  genotype_matrix,
  covariate_matrix,
  n_causal = 1000,
  n_epistatic = 20,
  heritability = 0.6,
  rho = 0.8,
  maxT = 100,
  scale_effects = 1,
  group_ratio = 1,
  maf_threshold = 0.01,
  seed = 67132,
  logLevel = "INFO",
  logFile = NULL
)
}
\arguments{
\item{genotype_matrix}{Genotype matrix with samples as rows, and SNPs as columns.}

\item{covariate_matrix}{Matrix with fixed effect covariates}

\item{n_causal}{Number of SNPs that are causal.}

\item{n_epistatic}{Number of SNPs that are epistatic}

\item{heritability}{Broad-sense heritability.}

\item{rho}{Proportion of heritability explained by additivity.}

\item{maxT}{Maximum time to be observed.}

\item{scale_effects}{Scale effects}

\item{group_ratio}{Ratio of sizes of groups that interact, e.g. a ratio 1:3 would be value 3.}

\item{maf_threshold}{is a float parameter defining the threshold for the minor allele frequency not included in causal SNPs.}

\item{seed}{Random seed for simulation.}

\item{logLevel}{is a string parameter defining the log level for the logging package.}

\item{logFile}{is a string parameter defining the name of the log file for the logging output.}
}
\value{
A list object containing the trait data, the genotype data, as well as the causal SNPs and summary statistics.
}
\description{
Simulate time to event data
}
