% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttemapit.R
\name{ttemapit}
\alias{ttemapit}
\title{The time-to-event Marginal Epistasis Test}
\usage{
ttemapit(
  fixed,
  data = parent.frame(),
  kins = NULL,
  id,
  tte,
  maxiter = 500,
  tol = 1e-05,
  taumin = 1e-05,
  taumax = 1e+05,
  tauregion = 10,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{fixed}{The formula for the covariates}

\item{data}{The dataframe with the traits and the covariates}

\item{kins}{The list of covariance matrices}

\item{id}{The string denoting the sample ID column in data}

\item{tte}{The string denoting the time to event column in data}

\item{maxiter}{Integer for the maximum number of iteratins in the optimization routine}

\item{tol}{Float for the convergence criterium}

\item{taumin}{Float that is minimum tau}

\item{taumax}{Float that is maximum tau}

\item{tauregion}{Float that is range of tau}

\item{verbose}{Boolean that toggles print outs.}
}
\value{
Named list with model parameters etc.
}
\description{
The time-to-event Marginal Epistasis Test
}
