\name{example}
\alias{example}
\title{Example dataset}
\description{
Example dataset for GMMAT.
}
\format{
  Contains the following objects:
  \describe{
    \item{pheno}{a data frame of 400 observations from a cross-sectional study with 5 variables: id, disease, trait, age and sex.}
    \item{pheno2}{a data frame of 2,000 observations from a longitudinal study with 400 individuals and 5 variables: id, y.repeated, y.trend, time and sex.}
    \item{GRM}{a genetic relationship matrix for 400 observations.}
  }
}


